web/requirements.txt
django-rest-swagger==2.2.0
django==2.2.8
djangorestframework==3.10.3
django-filter
psycopg2
django-import-export==1.2.0
pandas
numpy==1.17.4
tablib
biopython
uwsgi==2.0.15
ipython
